// src/types.ts
export type TBoardLocationIDs =
  | '9'
  | '10'
  | '11'
  | '12'
  | '13'
  | '17'
  | '18'
  | '19'
  | '20'
  | '21'
  | '25'
  | '26'
  | '27'
  | '28'
  | '29'

export type TBoardLocationDescriptions =
  | 'Forest 1'
  | 'Forest 2'
  | 'Forest 3'
  | 'Forest 4'
  | 'Forest 5'
  | 'Grassland 1'
  | 'Grassland 2'
  | 'Grassland 3'
  | 'Grassland 4'
  | 'Grassland 5'
  | 'Wetland 1'
  | 'Wetland 2'
  | 'Wetland 3'
  | 'Wetland 4'
  | 'Wetland 5'